+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [1.0.2] - 2019-08-28
+###Security
+- Updating dependencies
+
+## [1.0.1] - 2019-07-29
+###Security
+- Updating dependencies
+
 ## [1.0.0] - 2019-07-08
 ### Changed
 - Paths are now always strings.
@@ -106,6 +114,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [diffUpdate](docs/diffUpdate.md)
 - [intersection](docs/intersection.md)
 
+[1.0.2]: https://github.com/DarrenPaulWright/object-agent/compare/v1.0.1...1.0.2
+[1.0.1]: https://github.com/DarrenPaulWright/object-agent/compare/v1.0.0...1.0.1
 [1.0.0]: https://github.com/DarrenPaulWright/object-agent/compare/v0.6.0...1.0.0
 [0.6.0]: https://github.com/DarrenPaulWright/object-agent/compare/v0.5.3...0.6.0
 [0.5.3]: https://github.com/DarrenPaulWright/object-agent/compare/v0.5.2...0.5.3